refactor: split mcp server registration
Move MCP helpers, tool registration, resources, and prompts into focused modules while preserving the public server entrypoints and dynamic test patch behavior.
